+++ Runtime/Scripts/FSM/IStateMachineOwner.cs
@@ -3,5 +3,7 @@ namespace QuickDemo.FSM
 {
     public interface IStateMachineOwner
     {
+        bool IsFSMDebug { get; }
+        string FSMDebugLogPrefix { get; }
     }
 }